% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_delimKB}
\alias{read_delimKB}
\title{Read files of different format}
\usage{
read_delimKB(path)
}
\value{
A data frame
}
\description{
Technically this returns the data frame of the uploaded file of various format(csv, xlsx, txt).
To be noted that this function was firstly created by Krzysztof Bajdzienko(https://github.com/kbajdzienko).
}
\examples{
read_delimKB(path)
}
\seealso{
\link[readxl:reax_xlsx]{readxl::reax_xlsx},\link[readr:read_csv]{readr::read_csv} which this function wraps.
}
